Cosmetic changes in the Product and some in the Cart
Added the dto to get all products only with image name price gross and it was implemented in the product component in the all products view. Additionally, the sorts fa icons in the all products view were hidden. Transactional annotations are above in every method. That's why  annotation Transactional above class is not required